fix: Skip release prep steps if no changes to release (CORE-774)
Noticed while debugging why the fix for CORE-774 wasn't generating a
release after merging to main.  Determined this was because the script
looks specifically for conventional commit messages and if developers
don't realise that their fixes won't generate a release, and in fact
cause the release preparation to cryptically fail.